requirements.txt
cffi==1.13.2
cryptography==2.8
psycopg2-binary==2.8.3
pycparser==2.19
six==1.13.0
